+++ src/Concerto/TestBundle/Resources/R/forker.R
@@ -40,6 +40,7 @@ switch(ENV_CONCERTO_R_SESSION_STORAGE,
 concerto.log("starting forker listener")
 queue = c()
 unlink(paste0(ENV_CONCERTO_R_SESSION_FIFO_PATH, "*.fifo"))
+lastForcedGcTime = as.numeric(Sys.time())
 repeat {
     fpath = ""
     if(length(queue) == 0) {
@@ -74,7 +75,11 @@ repeat {
 
     if(is.null(response$rLogPath)) response$rLogPath = "/dev/null"
 
-    gcOutput = gc(F)
+    currentTime = as.numeric(Sys.time())
+    if(currentTime - lastForcedGcTime > 86400) {
+        gcOutput = gc(F)
+        lastForcedGcTime = currentTime
+    }
     mcparallel({
         if(ENV_CONCERTO_R_SESSION_LOG_LEVEL > 0) {
             sinkFile <- file(response$rLogPath, open = "at")